Remove bubble delimiter in quickfix window entries
The highlighting of these bubbles can't get resolved propperly. The
origin of this issue gets described in the previous commit. Basically
this colors the body of the bubble content different, but havea static
coloring for the delimiters. This just looks ugly. Unfortunately there
is no way to determine the correct color for each delimiter in this
context.
Thereby the remaining solutions are either to color all bubbles the
same, no matter the content, or color them differently, but don't have
delimiters. As the contents highlighting has an important effect, this
is the current go-for-solution.