Add a generic bubble creation function
The concept of "bubbeling" content gets more and more used as a visual
improvement. So far this produces a lot of duplicated code. Furthermore
was the coloring issue not solved properly yet. Only the coloring of
labels is implemented properly. For the other bubbles, a new utility
function got introduced. Unifying these both function is the target
of future improvements (atm they use incompatible color encoding). Note
that this function requires to provide a single highlight group only.
The new implementation gets used for reactions as well as for comment
states/properties.