Extend bubble factory to allow padding of content
This optional parameter is a nice to have that might be helpful for some
parts of the UI. It got introduced because the currently used reaction
unicode characters sometimes exceed the size of a single column.
Therefore they were cut off. Shifting the responsibility to the
declaration itself sound reasonable. But as it is a nice addition that
just works in general and make the UI appear more homogeneous (no single
reaction looks wider than the other), this approach seems the best for
the current situation.

For me the icon/symbols/unicode-chars (name them how you want) do not
all work. I'm using a NerdFont which is widely used by terminal lovers
who like fancy textual interfaces. But apparently some chars do not
work. Also that they alternate in their column width is a hint they do
not align for mono fonts. This should be tackled in future improvements.