Add icon to user bubbles with new factory
In the WebUI its is obvious that a possibly awkward collection of
letters actually revers to a user name. This is especially because there
is avatar next to the user. This is at least true for user names in
headers of objects like issue bodies or comments.
To improve the readability of the Octo UI, the bubbles content for a
user got preceded with a icon character that symbolized a user. Note
that we only bubble users for authors of issues, PRs or comments.
Thereby we align with the WebUI.

The creation of a bubble for a user got extracted into a new factory
function to remove duplicated code and simplify the higher level
function which render the whole content of an object. This also reduces
it to a single location where the icon gets defined.