Move highlight definitions from syntax to plugin
Setting the highlights just in the syntax file is too late. This got
undetected as I had custom definitions for these groups at my local
configuration. Anyways it is pretty uncommon to define new highlight
groups in a syntax file, but it is common to define them in the plugin
script. Thereby they are always available, without every to worry they
aren't. Also this should be no performance issue. Actually it is rather
a boost as they do not get (attempted) to redefine for each new buffer
that get the `octo_issue` filetype set.