Add Hover action to show pending comments on float
This commit adds support to show pending comments on a float window when
the cursor is idle on a line containing a comment. The float is
displayed in the alternative window. That is, if comment is on RIGHT
side, then float is shown on the LEFT side and viceversa.

Refactor of code to bring all review related code into reviews.lua
Now `_review_id`, `_review_comments` and `_review_files` are local
variables since they are only accessed from reviews.lua