Extend highlights as hierarchy for customization
During the discussions about the design of what to bubble, which colors
and icons, it has turned out that we need/should have a more options to
customize the appearance. Therefore more highlight groups get added for
many special cases. But most of them simply link to their "parent"
group. This allows users users to either customize every detail or
just making a few adjustments.

One special case for the highlighting was to "remove" bubbles for
users which are not the viewer by setting their highlight to the same as
normal. But so far this tweak was affecting all bubbles without the
special viewer attribute like reactions.

Another little improvement is the option to have a hint icon for
reactions the viewer has placed. This is either an alternative to the
different highlighting of the bubble or can be simply used as extension.
Finally the user can just "disable" both, so all reactions look 100% the
same.