Improve bubble delimiter highlight generation
Setting new highlight groups via the Lua API requires to maintain
highlight namespaces which need to be "activated" everywhere needed etc.
pp. This is quite cumbersome (for now) and it was requested to
remove/simplify this.
Determining the color of a highlight group still requires to use the Lua
API, as using Vim function calls does sometimes not work when the
plugin gets lazy loaded (groups are not defined as code runs to fast).
But for creating new highlight groups, we can now convert the integer
representations of the API and re-use the already existing highlight
module functions.