feature/review refactor (#137)
* Collect thread instead of comments

Before, we were adding review comments in `_review_comments` which mixed
thread and first comment details.
We now collect the thread and use `comments.nodes[1]` to access the first
comment.

The goal is to be able to show all thread comments in `Octo review
start|resume` and not only those from the pending review

* Add reaction groups to thread comments graphql queries

* Show thread comments buffer on cursorhold

* correct green color

* Major refactor

* Get rid of octo_issue and octo_reviewthread filetypes

* Update README

* Fix user search query

* Fix missing check for review_thread_map not available in issue buffers

* Fix thread folding

* Fix broken `Octo issue create`

* Do not show outdated thread in diff buffers

* Fix incorrect review thread buffer name

* Fixes arithmetic operation on foldcolumn

writers.lua:634: attempt to perform arithmetic on field 'foldcolumn' (a string value)

* Fix errors

* Rename qf related functions

* Implement #126

* Improve README regarding the review proceess