fix: lack of default branch name causes repo display issue
Fixes https://github.com/pwntester/octo.nvim/issues/519

When a repo doesn't have a default branch name (i.e. a totally empty
repo) then Octo gives an error when trying to view that repo.

This fixes that. Additionally, because when using the fzf-lua picker
since the repo is previewed the error can show up when that repository
is highlighted.