710 Bugfix escape char func (#772)
* Fix escape_char for quote

The `"` need no escaping. The old version, which was introduced in this
[commit](https://github.com/pwntester/octo.nvim/commit/580dbab5a73feea534c13d9c3a4ebd68a8721397#diff-5a72a6018b10915bef41b81020b59395527ba5b4b844d222b4e97192b5450565R369),
has unexpected behaviour of adding a `\` in the resulting strings.

This change fixes this issue.

* Add plenary test for escape_cahr()

Not sure about the style guide for plenary tests. So added it here.

Further, I think one should perhaps add more special cahracters to the
test for not changing characters, but not sure what the expected
coverage of the test should be.

* Fix stylua formatting

---------

Co-authored-by: Will Dean <57733339+wd60622@users.noreply.github.com>