fix: make comment ranges work (#807)
Previous to this comment highlighting several lines and running the
command "Octo comment add" would never result in a multi-line comment.

This is seemingly because as soon as you enter the command line with
':' Neovim is switched back into normal mode and the "v" and "."
attributes to `vim.fn.line` no longer register for the line range.

Instead, we can use the explicit start and end marks for a selected
range to capture the currently selected range despite being in normal
mode at time of command run.

Signed-off-by: ldelossa <louis.delos@gmail.com>
Co-authored-by: Will Dean <57733339+wd60622@users.noreply.github.com>