Fix broken `:Octo review commit` command workflow (#1077)
* Fix broken processing for review commit

* Fix review commit comment published on wrong line

This commit fixes the fact that when we posted a review comment for a
specific commit, it would be published on the wrong line in github. I
found out the reason for this was that we were using the *full* PR diff
and not the commit-specific diff, so this commit fixes that.

* Fix review commit comment misplaced across buffers

This commit fixes a bug with the `:Octo review commit` workflow where
whenever you added a comment on a "commit level", within octo.nvim's
file buffers, it would place signs for the comment across buffers (all
the different files changed within the commit around the same line), and
it would also not place the sign for the comment exactly where it should
be. This fixes that.

The bug was essentially that we didn't have precise checks in the plugin
on when we should be placing the signs. This will be apparent as you
review the code.

---------

Co-authored-by: Will Dean <57733339+williambdean@users.noreply.github.com>